Move several util methods to utils.js
